requirements.txt
pandas>=2.1.0
pymysql>=1.1.0
python-dotenv>=1.0.0
requests>=2.31.0
google-generativeai>=0.3.0
cryptography>=40.0.0
tabulate>=0.9.0
pytest>=7.0.0
streamlit>=1.26.0
